refactor: 메인화면 분리
